changed from link to navigate
why?:

Link and NavLink are mostly same thing.We use both of them to route pages.But the difference is when we use NavLink we get some advantages like we can design our navigation with active state.Because the NavLink component provides a active class inside it.So we can design our navigation when it is active and we can keep track the active pages.

useNavigate is a hook which returns a function to navigate.But to do this we need to call a navigate function and it declares that how it will work.